Fix error mark position for PatternParseException
PatternParseException.toDetailedString() return a String
with a mark to specify the error position in the pattern.
The mark takes place in the second line in the String
returned. Because PatternParseFailureAnalyzer.analyze
appended "Invalid mapping pattern detected:" at the
beginning of the returned String, the mark was not well
positioned.

Now, a "\n" is inserted after "Invalid mapping pattern detected:"
and the mark is well positioned

See gh-38944