Set `LOG_CORRELATION_PATTERN` only when correlation ID is expected
Prior to this commit, if `logging.pattern.correlation` has been set,
then the correlation ID is logged regardless of the flag
`logging.expect-correlation-id`. This is different from what happens
with the default correlation patterns.

This commit updates LoggingSystemProperties so that the system property
`LOG_CORRELATION_PATTERN` is set when `logging.expect-correlation-id`
is `true`. If it is `false` then the correlation ID won't be logged in
any cases.