Add link to Pulsar env var workaround
There is a section in the Pulsar client authentication reference guide
that warns users of the lack of relaxed binding for authentication
parameter map keys.

This lack of relaxed binding prevents users from setting these auth
parameters directly via env var as the casing is lost in translation.

The commit adds a link in this area of the reference guide to a
workaround in the Spring Pulsar framework reference guide.

See gh-39630