Do not overwrite attributes in ClientObservationConventionAdapter
Prior to this commit, the `ClientObservationConventionAdapter` would
overwrite a request builder attribute. This would happen when the
request is not fully built when the observation starts. At that point,
the tags are built for long task timers, but not for the actual metric.
This effectively overrides the correct value of the URI template in the
builder.

This commit removes this builder update which was invalid in the first
place.

Fixes gh-40330