Use ubuntu-latest GitHub Actions runner for Linux builds
A full build requires a lot of disk space. I believe this is mostly
due to the large number of Docker images that the tests utilize. By
default, the amount of free disk space on an ubuntu-latest runner
does not meet or needs. In this commit, we work around this using
the free-disk-space action which deletes things like the Android
SDK and Haskell installations that are both large and not needed by
our builds.

Closes gh-40705