Reply with HTTP 415 for unsupported GraphQL content-type
Prior to this commit, the configured GraphQL routes would reply with an
HTTP 404 status when a POST request is sent with an unsupported content
type, such as "text/plain". While such requests are not supported in the
first place, we should help developers and let them know that the
content type sent is the problem.

This commit configures new routes that reply with HTTP 415 "Unsupported
Media Type" for these cases.

Closes gh-41675