Don't pass -ea suffix into Gradle as the toolchain version
This commit is similar to what we had to do for Java 22-ea.

Given our actions have changed quite a bit, this commit replaces the
"java-distribution" parameter by a "java-early-access" parameter. When
set, it automatically switches the distribution to temurin as well as
applying the same handling of the Java version as in b8a6cab.

See gh-41698