Use early static registration of EventPublishingContextWrapper
Add `OpenTelemetryEventPublisherApplicationListener` which uses a single
static `ContextStorage` wrapper which gets applied as early as possible.
The static wrapper is then updated as beans come and go.

By adding the wrapper early, we hope to avoid silent failures which
can occur if the `ContextStorage` gets initialized before the wrapper
is added.

Closes gh-41439