Increase timeouts in ZipkinWebClientSenderTests
This is an attempt to stabilize the tests. They are currently flaky
due to a timeout that appears to occur because the mock web server isn't
responding quickly enough. A larger timeout will either confirm this if
the tests stabilize or indicate that the problem has a different cause
if they do not stabilize.

Closes gh-42375