Configure additional GraphQL schema files
Prior to this commit, the GraphQL auto-configuration would provide
configuration properties for specifying the locations and file
extensions of schema files to be scanned during startup.

This commit adds a new "spring.graphql.schema.additional-files" that
applications can use to point at particular files, like
"classpath:mylocation/schema.graphqls". These files will be parsed and
considered for the schema in addition to the scanned ones.

Closes gh-42792