GraphQL should not back off when GrapQlSource is present
Prior to this commit, the GraphQL auto-configuration that defines the
infrastructure beans for base support would only be active when:

* GraphQL schema files are detected in the configured locations
* or if GraphQlSourceBuilderCustomizer beans are present

This would allow some "code first" approaches, but not situations where
developers contribute their own `GraphQlSource`. This commit ensures
that the auto-configuration is processed even if the application only
contributes a custom `GraphQlSource` bean.

Closes gh-33096