Fix SSL file watcher for relative symlinks
Prior to this commit, a change in gh-43586 unlocked the support for
symlinks: instead of watching the link itself which might never change,
this would watch the target file which is likely to change.

This could break with an `IllegalStateException` in case the symlink is
using a path relative to the link itself.

This commit ensures that the target is resolved against the current
link path to avoid incorrect watch operations.

Fixes gh-43966